Move class ContinuousTransformBroadcaster into new header file
